Bugfix: Make it able to write more than 64k - 1 COFF relocations
If the relocation count is greater than 0xffff then it can't fit
in the NumberOfRelocations field of a section header. In order to
specify greater relocation count IMAGE_SCN_LNK_NRELOC_OVFL flag
should be added to Characteristics and the actual relocation count
should be written into VirtualAddress field of the first COFF
relocation.